Add custom decorators
Rename ModelDecorator as ModelDecoratorExt. It was a modified version of
the original ModelDecorator.
Create TosDecorator. It uses ModelDecorator’s DiagramDesigner and
PartBrowser. TosDecorator will be used as the custom Decorator for the
project. ModelDecoratorExt is used for the custom decorator. It’s to be
obsolete for future.
Add TestDecorator. It’s an exact copy of the latest ModelDecorator. It
will be used for test purposes.